Dataset from Richmond et. al 2017 investigating the association of BMI on insulin resistance

A two-sample summary level dataset, Richmond et al. (2017) https://dx.doi.org/10.1101/155739, containing 14 single nucleiodtide polymorphisms which have genotype-exposure associations (BMI) and genotype-outcome associations (insulin) with their respective standard errors.

bmi_insulin

Format

A data frame with 14 rows and 44 columns:

rsid

The genetic variant used as instruments

beta.exposure

The value for the phenotype-exposure association

beta.outcome

The value for the phenotype-outcome association

se.exposure

The value for the standard error of the phenotype-exposure association

se.outcome

The value for the standard error of the phenotype-outcome association

Details

bmi_insulin.

References

Richmond, R. et al., Investigating the role of insulin in increased adiposity: Bi-directional Mendelian randomization study. bioRxiv, 2017, https://dx.doi.org/10.1101/155739.