Bayesian inverse variance weighted model with a choice of prior distributions fitted using Stan

Bayesian inverse variance weighted model with a choice of prior distributions fitted using Stan.

mr_egger_stan(
  data,
  prior = 1,
  n.chains = 3,
  n.burn = 1000,
  n.iter = 5000,
  seed = 12345,
  rho = 0.5,
  ...
)

Arguments

data
A data of class mr_format.

prior
An integer for selecting the prior distributions;

1 selects a non-informative set of priors;

2 selects weakly informative priors;

3 selects a pseudo-horseshoe prior on the causal effect;

4 selects joint prior of the intercept and causal effect estimate.

n.chains
Numeric indicating the number of chains used in the HMC estimation in rstan, the default is 3 chains.

n.burn
Numeric indicating the burn-in period of the Bayesian HMC estimation. The default is 1000 samples.

n.iter
Numeric indicating the number of iterations in the Bayesian HMC estimation. The default is 5000 iterations.

seed
Numeric indicating the random number seed. The default is 12345.

rho
Numeric indicating the correlation coefficient input into the joint prior distribution. The default is 0.5.

...
Additional arguments passed through to rstan::sampling().

Value

An object of class stanfit.

References

Bowden J, Davey Smith G, Burgess S. Mendelian randomization with invalid instruments: effect estimation and bias detection through Egger regression. International Journal of Epidemiology, 2015, 44, 2, 512-525. http://dx.doi.org/10.1093/ije/dyv080.

Stan Development Team (2020). "RStan: the R interface to Stan." R package version 2.19.3, http://mc-stan.org/.

Examples

# \donttest{
# Note we recommend setting n.burn and n.iter to larger values
egger_fit <- mr_egger_stan(bmi_insulin, n.burn = 500, n.iter = 1000)

#> 
#> SAMPLING FOR MODEL 'mregger' NOW (CHAIN 1).
#> Chain 1: Rejecting initial value:
#> Chain 1:   Error evaluating the log probability at the initial value.
#> Chain 1: Exception: normal_lpdf: Scale parameter is -1.87978, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 1: Rejecting initial value:
#> Chain 1:   Error evaluating the log probability at the initial value.
#> Chain 1: Exception: normal_lpdf: Scale parameter is -0.223268, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 1: Rejecting initial value:
#> Chain 1:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 1:   Stan can't start sampling from this initial value.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.489 seconds (Warm-up)
#> Chain 1:                0.188 seconds (Sampling)
#> Chain 1:                0.677 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'mregger' NOW (CHAIN 2).
#> Chain 2: Rejecting initial value:
#> Chain 2:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 2:   Stan can't start sampling from this initial value.
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -1.18416, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -0.677049, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -1.50931, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -1.94189, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -1.22206, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: Rejecting initial value:
#> Chain 2:   Error evaluating the log probability at the initial value.
#> Chain 2: Exception: normal_lpdf: Scale parameter is -0.0407303, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 2: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 2: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 2: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 2: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.35 seconds (Warm-up)
#> Chain 2:                0.163 seconds (Sampling)
#> Chain 2:                0.513 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'mregger' NOW (CHAIN 3).
#> Chain 3: Rejecting initial value:
#> Chain 3:   Error evaluating the log probability at the initial value.
#> Chain 3: Exception: normal_lpdf: Scale parameter is -1.11539, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 3: Rejecting initial value:
#> Chain 3:   Error evaluating the log probability at the initial value.
#> Chain 3: Exception: normal_lpdf: Scale parameter is -1.17308, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 3: Rejecting initial value:
#> Chain 3:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 3:   Stan can't start sampling from this initial value.
#> Chain 3: Rejecting initial value:
#> Chain 3:   Error evaluating the log probability at the initial value.
#> Chain 3: Exception: normal_lpdf: Scale parameter is -1.11347, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 3: Rejecting initial value:
#> Chain 3:   Error evaluating the log probability at the initial value.
#> Chain 3: Exception: normal_lpdf: Scale parameter is -1.60474, but must be > 0!  (in 'model_mregger' at line 65)
#> 
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 3: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 3: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 3: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 3: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.528 seconds (Warm-up)
#> Chain 3:                0.192 seconds (Sampling)
#> Chain 3:                0.72 seconds (Total)
#> Chain 3: 

#> Warning: There were 284 divergent transitions after warmup. Increasing adapt_delta above 0.999 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

print(egger_fit)

#> Inference for Stan model: mregger.
#> 3 chains, each with iter=1000; warmup=500; thin=1; 
#> post-warmup draws per chain=500, total post-warmup draws=1500.
#> 
#>             mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> intercept  -0.05    0.00 0.04  -0.12  -0.08  -0.06  -0.03   0.02   323 1.01
#> estimate    3.78    0.12 2.20  -0.47   2.25   3.91   5.30   8.05   345 1.01
#> sigma       7.65    0.08 1.17   5.58   6.77   7.56   8.55   9.87   213 1.01
#> lp__      -35.43    0.07 1.14 -38.52 -35.93 -35.17 -34.63 -34.10   293 1.01
#> 
#> Samples were drawn using NUTS(diag_e) at Fri May 29 10:08:54 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }