Bayesian inverse variance weighted model with a choice of prior distributions fitted using RStan.

Bayesian inverse variance weighted model with a choice of prior distributions fitted using RStan

mr_radialegger_stan(
  data,
  prior = 1,
  n.chains = 3,
  n.burn = 1000,
  n.iter = 5000,
  rho = 0.5,
  seed = 12345,
  ...
)

Arguments

data
A data of class mr_format.

prior
An integer for selecting the prior distributions;

1 selects a non-informative set of priors;

2 selects weakly informative priors;

3 selects a pseudo-horseshoe prior on the causal effect;

4 selects joint prior of the intercept and causal effect estimate.

n.chains
Numeric indicating the number of chains used in the HMC estimation in rstan, the default is 3 chains.

n.burn
Numeric indicating the burn-in period of the Bayesian HMC estimation. The default is 1000 samples.

n.iter
Numeric indicating the number of iterations in the Bayesian HMC estimation. The default is 5000 iterations.

rho
Numeric indicating the correlation coefficient input into the joint prior distribution. The default is 0.5.

seed
Numeric indicating the random number seed. The default is 12345.

...
Additional arguments passed through to rstan::sampling().

Value

An object of class stanfit.

References

Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278. https://doi.org/10.1093/ije/dyy101.

Stan Development Team (2020). "RStan: the R interface to Stan." R package version 2.19.3, http://mc-stan.org/.

Examples

# \donttest{
# Note we recommend setting n.burn and n.iter to larger values
radegger_fit <- mr_radialegger_stan(bmi_insulin, n.burn = 500, n.iter = 1000)

#> 
#> SAMPLING FOR MODEL 'mrradialegger' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.185 seconds (Warm-up)
#> Chain 1:                0.171 seconds (Sampling)
#> Chain 1:                0.356 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'mrradialegger' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 2: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 2: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 2: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 2: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.184 seconds (Warm-up)
#> Chain 2:                0.12 seconds (Sampling)
#> Chain 2:                0.304 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'mrradialegger' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 3: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 3: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 3: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 3: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.161 seconds (Warm-up)
#> Chain 3:                0.154 seconds (Sampling)
#> Chain 3:                0.315 seconds (Total)
#> Chain 3: 

#> Warning: There were 119 divergent transitions after warmup. Increasing adapt_delta above 0.999 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

print(radegger_fit)

#> Inference for Stan model: mrradialegger.
#> 3 chains, each with iter=1000; warmup=500; thin=1; 
#> post-warmup draws per chain=500, total post-warmup draws=1500.
#> 
#>             mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> intercept -30.08    0.63 12.86 -54.87 -38.68 -30.06 -21.04  -6.30   412 1.00
#> estimate    6.28    0.12  2.47   1.73   4.64   6.27   7.87  11.28   421 1.00
#> sigma       6.94    0.07  1.25   4.82   5.97   6.86   7.83   9.35   367 1.01
#> lp__      -33.91    0.06  1.12 -36.77 -34.54 -33.72 -33.03 -32.47   374 1.00
#> 
#> Samples were drawn using NUTS(diag_e) at Fri May 29 10:08:58 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }